v0.3: fix transparent pixels in various Tiles
The empty pixels are introduced by the XBR algorithm. It is probably a bug,
but the source code of XBR looks hard to understand, so as a workaround I
create a horizontally mirrored image of the Tiles, magnify that as well,
mirror it back to original, then "merge" these 2 images. Merge means use
the pixel with higher alpha in each (x,y) coordinate.